+++ apps/extension/src/entrypoints/options/app-sidebar/index.tsx
@@ -1,7 +1,7 @@
 import { i18n } from '#imports'
 import { Icon } from '@iconify/react'
 import { Link } from 'react-router'
-import readFrogLogo from '@/assets/icons/read-frog.png'
+// import readFrogLogo from '@/assets/icons/logo.png'
 import {
   Sidebar,
   SidebarContent,
@@ -21,7 +21,7 @@ export function AppSidebar() {
     <Sidebar collapsible="icon">
       <SidebarHeader className="group-data-[state=expanded]:px-5 group-data-[state=expanded]:pt-4 transition-all">
         <a href="https://readfrog.app" className="flex items-center gap-2">
-          <img src={readFrogLogo} alt="Logo" className="h-8 w-8 shrink-0" />
+          {/* <img src={readFrogLogo} alt="Logo" className="h-8 w-8 shrink-0" /> */}
           <span className="text-md font-bold overflow-hidden truncate">{i18n.t('name')}</span>
           <span className="text-xs text-muted-foreground overflow-hidden truncate">
             {`v${version}`}